requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
google-auth==2.47.0
google-auth-oauthlib==1.2.4
gspread==6.2.1
gunicorn==23.0.0
idna==3.11
oauthlib==3.3.1
packaging==25.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.11.0
